# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexus Platform API",
    "description": "Investor & Entrepreneur Collaboration Platform — REST API documentation",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000/api", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/profiles/entrepreneur/{userId}": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Get entrepreneur profile",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Entrepreneur profile" } }
      },
      "put": {
        "tags": ["Profiles"],
        "summary": "Update entrepreneur profile",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Profile updated" } }
      }
    },
    "/profiles/investor/{userId}": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Get investor profile",
        "parameters": [{ "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Investor profile" } }
      }
    },
...
  },
...
}